backend/requirements.txt
absl-py==2.0.0
amqp==5.2.0
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto3==1.28.67
botocore==1.31.67
cachetools==5.3.1
celery==5.3.4
certifi==2022.12.7
channels==4.0.0
channels-redis==4.1.0
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
contourpy==1.1.1
cron-descriptor==1.4.0
cycler==0.10.0
Django==4.2.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
exceptiongroup==1.1.3
filelock==3.12.4
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.9.2
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.1.0
grpcio==1.59.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.18.0
humanfriendly==10.0
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.5
kombu==5.3.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2
numpy==1.26.1
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.16.1
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pyparsing==2.4.7
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.7
rsa==4.9
s3transfer==0.7.0
sahi==0.11.14
scipy==1.11.3
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sqlparse==0.4.4
supervision==0.16.0
sympy==1.12
tensorboard==2.15.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.202
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.9
websockets==12.0
Werkzeug==3.0.0
yolov5==7.0.13
yolov8==0.0.2
zipp==3.17.0
